Bug 708605: Cope with /AS being an indirection in annotations.
Usually we expect the annotation /AS entry to be a direct name,
such as:   /AS /Off

As part of annotation resynthesis, we look that up, and set that key
to a value. So:   /Off <something>

But in the example given in this bug, /AS is an indirect reference
to a name:

 10 0 R
 /Off
 endobj

and then:   /AS 10 0 R

This trips us up as we attempt to use the indirect reference as
the key to write with.

Rather than opening up our API by allowing indirect references to
names to be used as keys (which would probably mask problems
elsewhere), just resolve AS once read in this particular case.

Thanks to Sebastian for spotting other places that do the same
thing.