mudraw: Report errors when ignoring them.
Running mutool draw -iDst pdfium/testing/resources/bug_1324189.pdf

That PDF is broken, but mutool draw handles error reporting incorrectly and
prints a message about an unhandled exception.

The option -i means that errors should be ignored. Unfortunately the mudraw
code does not report ignored errors, but leaves them on the error stack. When
the next exception is thrown, the error stack is found to not be empty and the
message about an unhandled exception is printed.

With this commit the errors are reported (and thus the error stack emptied)
before printing a warning about ignoring the errors.