Fix HTML pagebreaks for left and right pages.
The calculations for these in the normal case were distinctly
screwy. They certainly broke when used with stories, where the
rects can be different height, so the calculation of whether
we were on left or right pages was untrustworthy.

For non-restarting contexts (i.e. non-Story use), we now calculate
the values as before (with less screwy calculations) using
page_h.

For restarting contexts, we pass in whether we are a left or
right page (and alternate this automatically for each rect).
Our first rect is always a 'right' page.

Potentially in future we can add a functionality like
fz_story_set_handedness() to say whether the next place should
be to a left or right rectangle if required.

Also, we add left/right pagebreak tests to storytest.