Change INSTALL to use separate submodule initialization step.
When cloning recursively, git will fetch all tags and branches of all
submodules; but when initializing submodules after cloning it will only
fetch the required HEAD branch.

The way git clone handles recurse-submodules depends on the flags passed
to the initial clone -- so unless you use --single-branch git will also
clone all the tags and branches of all submodules; which is something
we don't want.

Using a second explicit submodule initialization step is much faster
and downloads much less data.

NOTE: We have set up an "artifex" branch for each submodule which is
used as the branch for the initial submodule clone.