Bug 708643: Avoid cached tiles conflicting between documents.
We update the device interface so that we can pass a document
id as well as a tile id. The draw device can then store cached
tiles with a document id reference, meaning we will no longer
accidentally match tiles from different documents.

Every document gets a context-unique id generated at creation
time from the context.

Credit is due too to Sebastian and Tor for the ideas in here.