Tweak handling of ActualText in Structured Text device.
When ActualText is applied to an image, we rely on the image
being sent to the device for us to know where it should go on
the page.

We were suppressing all images when the device was not collecting
them, so we were getting warnings about ActualTexts with no
contents.

Accordingly, we now only suppress image decoding when we are
also ignoring actualtext.

Also, don't warn about ActualTexts with no content if we are using
the 'ignore actualtext' flag.

Thanks to Sebastian for spotting the problem.