Fix whitespace bug in searching.
To reproduce: Open page 554 of pdf_reference17.pdf in mupdf-gl.
Search for " matte ". Hit 'n' to get the next page. Assertion
given.

While searching we have to keep the pointers into the haystacks
in sync with the pointers into the stext. If the point we are
remembering starts with whitespace, and that's implicit in the
stext rather than actually being there, our logic gets confused.

We therefore have to have some code that copes with this case
when we restart searching.