Fix overlap checking in add_range for 'many' ranges.
Ranges with 'many' set can only be added to the tree by
calls to add_range() from add_mrange. In such cases, they
always have low == high.

At first glance it looks like more such ranges can be added
from the recursive call to add_range within add_range, but
closer analysis shows that this can never give rise to calls
with many=1.

In order for the recursive case to call with many=1, there
must be an entry in the tree that has low < high, and many=1.
Given that no other call can insert such a node, this call
can't insert such a node either.

The logic for case 1 and 2 were incorrect in the presence of
a 'many' node. The correct thing to do is just to remove the
existing node, and take the next one. The logic is updated
here to do that.

This fixes oss-fuzz issue 428760786. Thanks to Sebastian for
identifying the problem!