Bug 708660: Fix mistake in renumbering objects when saving.
When we save with renumbering, we have a special case for coping with
objects such as:

47 0 obj
46 0 R
endobj

This occurs very rarely in real files, and so we hadn't noticed that
our code was completely wrong.

We were rewriting this to be:

47 0 obj
47 0 R
endobj

when we should have been rewriting it to be:

47 0 obj
<X> 0 R
endobj

where <X> is whatever number object 46 was being renumbered to be.