Bug 708268: Do not override default layer state visibility.
Previously the code ignored the layer state in preference to the
Usage setting in the OCG dictionary. That means that user actions
that affect the state are ignored.

Hide the optional content layer if either the state or Usage are
set to OFF, only show it if both state and Usage are set to ON.