Fix PDF device's handling of stroked paths.
The PDF device has a cunning optimisation in it to spot paths
that are really axis aligned rectangles. Unfortunately,
this triggers for some stroked paths, such as:

 0 0 m 100 0 l

where it really shouldn't, as it results in such lines being
wrong when they are dashed.

Accordingly, we fix this by adding a utility function that
tells us whether all segments of a path are closed.