Improve alignment of table cells.
A row of cells, all with different paddings and border widths
will have the top and bottoms of the border aligned for all the cells.

Effectively this means there is extra 'secret' padding added
to the top and bottom of table cells.

We add support for this by reusing the 'margin' fields that are
explicitly unused for cells.