Fix Table regressions.
Malc supplied examples that show different problems.

Cut down, they resolve to the following simple tables:

1) A table with no cells in the first row:

 <table>
  <tbody>
   <tr><td display="hidden">A</td></tr>
   <tr>B</tr><tr>C</tr>
  </tbody>
 </table>

This was causing problems with cells in the top row not
being marked as spanned, so we tried to access cell->box
which was NULL. This is fixed by ensuring that we set
spanned=1 on all cells 'exposed' by us growing the table.

2) A table with a column with no independent cells in it.

 <table>
  <tbody>
   <tr><td colspan="3">A</td></tr>
   <tr>B</tr><tr>C</tr>
  </tbody>
 </table>

Here, column 3 only exists because cell A spans across into
it. This was tripping an assert.

We fix this by updating the column width finder to reset such
'empty' columns to having an unfixed 0 width.

3) A table with missing cells:

 <table>
  <tbody>
   <tr><td rowspan="2">A</td><td>B</td></tr>
  </tbody>
 </table>

Here we are confused by the cell under 'B' not being spanned
but also not having a 'box' defined. Just spot box being NULL
and continue.