Bug 708690: When encountering unbalanced q/Q operators, use default gstate.
When the pdf_color_processor is allocated a default graphics state is
also allocated and initialized. When a Q operator was encountered in
pdf_color_Q() the graphics state would be dropped and freed, even the default
graphics state!

Later when a q operator was encountered in pdf_color_q(), a NULL dereference
would occur because that code assumes that a graphics state is always present.
The same is true for many other operators, e.g. pdf_color_fstar().

This commit changes pdf_color_Q() such that it if it pops the last graphics
state off of the stack, it will repopulate it with a new default graphics
state. This means that any code relying on a graphics state being present
will still work.