Bug 708615: Guard against gstate being moved.
This is an extension of bug 703151 where pdf_show_char()
temporarily gsaving/grestoring due to reallocation may cause the
gstate block to move. The gstate will still be at the same depth
(gtop value) when the caller is returned to, but the pr->gstate
pointer may be a different value which causes gstate pointers
held by the caller to become stale.

Bug 703151 fixed the issue in show_string(), by updating the
locally held gstate pointer to that from the pdf_run_processor
plus gtop.

The same problem also exists one level higher in pdf_show_text().
It also holds a gstate pointer while calling show_string().

Similar to the fix for 703151, the fix here is to update the
locally held gstate pointer in pdf_show_text() after the call to
show_string().