Bug 708696: Fix heap buffer overflow in table processing.
During development fixup_collapsed_cell_bottoms_for_row was called
with a box pointer for the row. Late on it changed to being called
with a box pointer for the entire table.

The code was not changed to properly account for this, resulting in
us walking the structure off-by-one level.