Bug 708617: Restrict directories from which mupdf loads DLLs.
Many thanks to Neil Mark Ochea for the report and discussion that
lead to this commit.

The Win32 build of MuPDF fails to restrict the directories from
which DLLs can be loaded, allowing a malicious 'fake' DLL to be
loaded, which potentially has access implications.

The fix, implemented here, is to ensure that we only load DLLs
from the system directory.