Bug 708511: Harden fz_warp_pixmap against overflows.
Thanks to Amir Mohammad Jahangirzad for the report that lead
to this.

With sufficiently large pixmaps, we could cause overflows in some
of the calculations here. The code is designed to work with
pixmaps that are small enough that this should never be a problem,
but it doesn't hurt to protect the code a bit.

We rearrange some expressions so that they won't overflow if int
is smaller than size_t; often this is done by putting a size_t
as the first thing to multiply.