mutool draw: Exit after message about multiple threads but no band height.
Previously all the threads were created, but never asked to render anything.
Instead the document was rendered using single thread rendering. This consumes
unnecessary resources.

An alternative is to print a warning without creating multiple threads, but
fall back to single thread rendering. That rendering does not match the
requested configuration by the options passed to mutool draw, and the
rendering might be unexpectedly slow (since multiple threads were expected
to be used).

It is better to error out early and the user to understand the configuration
istake and either set a band height, or remove the requested number of threads.