Bug 708737: mutool draw: Delay rendering thread creation and limit to number of bands.
Previously if the user specified -T 10000 then 10000 threads would
be created, but if the maximum number of bands for any page in the
document was 10, then no more than 10 threads would be used.

Unfortunately mutool draw doesn't know the page sizes until it tries
to render each page, which means that the number of bands per page
is similarly unknown.

This commit delays creation of rendering threads until the number of
bands for a page is known. At this point up additional threads are
started until the total number of threads equal the number of bands
for the current page. The consequence of this is that if pages 1-10
have 12 bands each, then 12 threads are created when rendering page
1, then those threads are reused for pages 2-10. If then page 11
has 20 bands, just before rendering page 11, an additional 8 threads
are created. If one or more of those threads cannot be successfully
created, then the entire process fails with an error message.