mutool draw: Print error message if banding is enabled without threads.
Enabling banding without at least one rendering thread does not make sense.
Currently mutool draw falls back to single thread rendering, which may does
match the requested configuration through mutool draw options.

It is better to error out early and let the user choose different options,
either by setting a number of threads or disable banding.

Also move the check for a negative band height to the same area of code that
validates the configuration of number of threads and band height. And warn
if the number of threads is negative.