Only set resynth_required when there are annotations present.
If there are no annotations, and we set the flag, we'll never clear the
flag and end up not being able to save the PDF due to the check for saving
a document with a dirty resynth_required flag.

Example: mupdf-gl create a redaction and apply it. After redaction there
are no annotations left, but since mupdf-gl reloads the page immediately
we end up with a document we can't save.