Treat reserved glyph name indices as .notdef in font subsetting of post2 tables.
The prior commit fixed an ASAN complaint while subsetting fonts in
tests_private/pdf/sumatra/multiply_luminosity_masks_ignored.pdf

The root cause was that in the post2 table of one of the used fonts, the glyph
name index was 0xffff, but there were not this many strings in the post2 table
causing an out of bounds access. The introduced fix was to throw an exception
if the glyph name index caused the code to step out of bounds.

This commit considers whether glyph name index 65535 is indeed a valid glyph
name index, and if so, how to treat it.

The Apple TrueType reference manual lists name indices [32768, 65535] as
reserved for future use in post2 tables. The Microsoft OpenType Specification
1.9.1 does not mention any reserved indices for these tables.

The conservative choice is to treat glyph name indices in the range [32768,
65535] as reserved for TrueType fonts. The way this commits decides to handle
reserved glyph names indices is to treat them as if they were set to 0, i.e.
the macroman glyph .notdef. Doing so allows MuPDF to successfully subset the
fonts in this file.