Throw exception on out of range glyph name indices in post2 table during font subsetting.
Running font subsetting on
tests_private/pdf/sumatra/multiply_luminosity_masks_ignored.pdf
causes an ASAN complaint.

The reason is that in the post2 table for one of the used fonts, the glyph name
index for is 0xffff, which indicates that the name of the glyph is stored in
post table string number 65535. The code then tries to locate this string to
check whether it is one of the macroman strings (these can be referenced more
efficiently by a glyph name index 0-257).

While iterating to locate post table string number 65535 the code does not
ensure that the location resides within the table size, therefore attempting
to read beyond the end of the post table causing the ASAN complaint.

This commit adds validation that the string location must be within the post
table size. If not, an exception is throw.