Update pdf stream compression bomb detection to account better for images.
Images (in particular, "single color" images) can compress much
much better than our compression bomb detection likes.

These are coped with fine during normal decompression and use, as
we know we are decompressing images, and so can predict the 'worst
case' size of the stream much better.

So, here we update the stream loading code to spot images and
(in the absence of already having been given a worst case value)
to guess it from the declared sizes.