SText device: Tweak vector detection with clipping
Non-rectangular paths were failing to be clipped properly.

The background to this is that I was doing text extraction on a page
and being told that there were vectors that completely covered it.

On looking into it, I found that there were various printers marks
surrounding the page, none of which actually cover the page. Because
these were done as multi segment (non-rectangular) paths, the bbox
for each path intersected the page, even though none of the bboxes
for the individual segments did.

This code improves it, but can still fall foul of some
edge cases.

For example:

 0 0 300 300 re S

with a clipping rectangle of (100,100)->(200,200) will not be
completely clipped away. To do it properly, we'd need to at
least approximate flattening beziers.

This will do for now.