Table Hunt; tweak reinforcement of table edges
When considering grid lines, we can use the end points of
vectors that look plausibly like grid lines to "reinforce"
our idea of where divisions between cells are.

Never accept this reinforcement for the outside edges of
the table, as this can only pull the edges in, and the edges
have been carefully calculated to include all the content. If
we move them inwards, it causes problems in later steps.