Bug 708804: Handle lines without any table cells in them.
The epub file contains:

  <table style="border-collapse:collapse">
    <tr/>
    <tr>
      <td></td>
    </tr>
  </table>

This causes the table hunting code to call collapse_table_borders()
which operates on a cell grid where all of the cells in the first
row lack a corresponding html box.

Several of the for-statements in collapse_table_borders() omits a
check for if the box is NULL before dereferencing check->box.
The first for-statement does a check however, and this commit adds
the same to the remaining for-statements.