Bug 708803: Always initialize svg parser state.
The svg file from the bug is constructed with several groups that
causes the svg parser to run out of exception stack space in
svg_run_g().

This means that local_state is never initialized, then fz_try()
bumps into the exception stack limit, without processing the rest
of the fz_try()-block that containing a call to svg_begin_state()
the fz_always()-block is run. In fz_always() there is a call to
svg_end_state() which processes the uninitialized local_state. This
leads to the stroke state in local_state being dropped which causes
a problem in ASAN/valgrind.

By moving the svg_begin_state() call outside fz_try(), this issue
can be avoided. The alternative would be to always initialize
local_state with zeroes, but then is a performance penalty for every
call for the svg parser in the normal case.