Bug 708838: Do not resolve Last object in outline node when comparing them.
The PDF file from the bug report has among its outline nodes PDF
objects below. When they are processed by pdf_test_outline() it
will look at 5997 0 R and compare its Last reference, 6001 0 R
with the actual last object in the list of Next references
starting with 6002 0 R and ending with 6100 0 R.

Because pdf_test_outline() uses pdf_objcmp_resolved() when
comparing the referenced Last and the actual Last outline node
and these two objects are identical no repair of the Last
reference in 5997 0 R will occur. Due to incorrect Parent
reference the Parent reference in 6100 0 R will be updated
however. Had this comparing been done using pdf_objcmp() instead,
then 6001 0 R and 6100 0 R would have been considered different
and the parent's Last reference would have been updated.

Later on pdf_test_outline() is called again, now with fixed ==
NULL which causes exceptions if any errors in the outline are
encountered. This does happen because 6001 0 R and 6100 0 R are
now different (because the differing Parent pointers). Which
means that the entire outline tree is discarded.

    5997 0 obj
    <<
      /First 5999 0 R
      /Count -2
      /Prev 6000 0 R
      /Last 6001 0 R
      /Title (2 References)
      /Next 6002 0 R
      /Parent 6003 0 R
      /Dest [ 456 0 R /XYZ 68 784 null ]
    >>
    endobj
    5999 0 obj
    <<
      /Dest [ 456 0 R /XYZ 68 753 null ]
      /Count 0
      /Title (2.1 Normative references)
      /Next 6100 0 R
      /Parent 6099 0 R
    >>
    endobj
    6100 0 obj
    <<
      /Prev 6097 0 R
      /Count 0
      /Title (2.2 Informative references)
      /Dest [ 456 0 R /XYZ 68 568 null ]
      /Parent 6098 0 R
    >>
    endobj
    6001 0 obj
    <<
      /Prev 6097 0 R
      /Count 0
      /Title (2.2 Informative references)
      /Dest [ 456 0 R /XYZ 68 568 null ]
      /Parent 6098 0 R
    >>
    endobj

git blame reveals that commit 019a63ec9fc430161b4756ed39b398eee4498a03
changed from pdf_objcmp() to pdf_objcmp_resolve(), but it appears
unconnected to the rest of the commit and the commit message. The
best bet is that this was accidentally included as part of the
commit and not discovered in code review.

Therefore this commit just reverts back to calling pdf_objcmp()
when comparing the Last reference, and that fixes the reported
PDF.