Bug 708836: Remove (stale) DecodeParms when deflating streams.
When we compress streams on output, if the previous stream has a bogus
DecodeParms (even if it has no Filter), this must be removed since
it won't match the new filters!

We take care not to compress already compressed streams (by checking
for an existing Filter). In the example file, the stream was not
compressed but it did have an unused DecodeParms which we did not
remove and which causes errors when trying to read the resulting file.