Bug 708825: Ensure piped binary data on stdout survives on Windows.
stdout (and indeed stderr) are not set to be binary streams by
default. For some (most) filetypes, this will cause corruption.

Previously, we've correctly set stdout to be binary where required
with direct code in mudraw.c, but this stopped working when we
moved to internalise our stdout/stderr handling in
fz_new_output_with_path.

Here we tweak it so that outputs created by fz_new_output_with_path
that happen to use stdout/stderr on windows will set these to be
binary automatically.

This means that piped data, like file data will always be generated
in binary mode. In theory this might annoy some people because
their editors expect CRLF terminated text files, but we've been
getting this "wrong" for years, and we prefer the idea that you
get identical files on all platforms.

Also, remove the code that sets the quiet flag if we are piping
to stdout. People should use -q instead if they want this.
Warnings are printed to stderr, not stdout, so shouldn't affect
piped data anyway.