Bug 708856: Swap color profiles in use during softmask creation.
Add an alternative set of color profiles; ps_gray, ps_rgb, ps_cmyk.

Update the draw device so that we signal whether we are in a
softmask creation phase or not to the color conversions
routines. We do this by using a bit in the rendering intent
byte.

If the color conversion routines are called with this bit
set, then we use these new icc profiles in preference to
other gray/rgb/cmyk ones.

This matches Ghostscript's behaviour.