Remove unnecessary NEON build flag, implicitly enabled for arm64-v8a.
Android builds, for arm64-v8a, set HAVE_NEON. The makefiles detect this,
and set appropriate flags for Tesseract. We used to set -mneon, but since
commit 63a97960af65aacab4c90b39290b249af49dc978 we updated it to set -mfpu=neon
as that is what modern compilers expect.

The MuPDF sample Android apps starting with 1.26.5a and later were built using
NDK 28.2.13676358 (includes clang 19.0.1) where the new -mfpu-flag worked well.

SmartOffice is built using NDK 29 (includes clang 21.0.0), where clang gets
confused when -mfpu= is fed to the linker because it is a compilation only
flag. The linker doesn't ignore the flag as it should and emits an error.

NEON is always enabled for aarch64 as mentioned in tesseract commit
d7c071197e771263e923b731af05302a7e68efbd and checks in
tesseract/src/arch/simddetect.cpp: "ARMv8 always has NEON".

Therefore we do not need to set any particular flag to enable NEON support and
so this commits removes setting -mfpu in Makelists. This works well when
building both the MuPDF sample apps for Android and SmartOffice.

Thanks to Robin and Alan for figuring this out!