Restore default output for mutool draw being stdout.
Prior to commit 7c77ee540d5c4a93b6c661d26e82b3bf76f34071, calling:

 mutool draw -F trace in.pdf

we'd see output on stdout. Since that commit it was sent on to a
file called '-'.

Also, '-o -' should have sent stuff to stdout, but instead it
went to a file called '-'.

Here we reintroduce the logic to map '-' to stdout, and to use
stdout as the default.