jni: Print all exceptions when failing to load native shared library.
MuPDF tries loading the JNI shared library under different names in this order:

 * libmupdf_java.so
 * libmupdf_java64.so
 * libmupdf_java32.so

If the first fails to load, then second is attempted, and so on. The problem is
if an application is expected to contains libmupdf_java.so, but it for some reason
fails to load. Then trying to laod libmupdf_java64.so and libmupdf_java32.so will
fail because those files are likely not present.

In this case the exception for loading libmupdf_java32.so is the only exception
printed to the log, which is unfortunate because the exception for why
libmupdf_java.so failed to load is more interesting.

This commit prints the exception stacktraces of all three exceptions if all three
attempts load the JNI shared library fails. This provides more information when
trying to understand why the library failed to load.