Bug 702012: Increase precision of matrix inversion.
The example file has a tiling pattern with a huge BBox and tiny Matrix.
When we were inverting this we were running into FLT_EPSILON precision
issues.

Use doubles to invert matrixes like in fz_try_invert_matrix.

Also return a degenerate (singularity) matrix when inversion fails,
so it's easier to spot the cause for similar problems in the future.