Bug 697651: Add a pdf_check_document function to manually check and repair.
Like "fsck" but on the open PDF file.

We need to call this before applying any editing passes in mutool clean;
or these edits will be lost when a repair is triggered.

This does not cause any performance regression, because the
pdf_save_document step is also doing this check in the prepass.
The new public function also remembers whether it's been called,
so we don't have to do the work again.