Bug 708878: Fix overprint problems with disabled spots.
Various different problems contribute to this bug, all down
to bits of code not responding to spots being disabled properly.

In general, the code assumed that separation n will appear
as spot plane n in the pixmap, when actually separation n
will appear as spot plane n-x, where x is the number of
disabled separations before n.

We fix this in the overprint setup code, in the PSD writing code,
and in the separation color handling code.

Also, add some more comments, and factor out a common loop
to make stuff simpler too.