Add fz_segment_stext_rect.
This is a call that allows us to run segmentation analysis on
just a small area of the page.

The primary change in this commit is really the generalisation
of the stext walking code used in the segmentation code.

Previously, we assumed that segmentation was being done on
'unstructured' stext data (i.e. there were no structure tags,
and it was just a list of blocks, rather than being a tree).

In order to cope with this being run on stext with structure
tags present, we need to modify the routines to recurse through
the structure. A position within the tree is therefore represented
not just with a block pointer, but also a parent pointer (which
tells us both which 'run' within the tree we are in, and how to get
back to the parent (and grandparent etc)).

Accordingly, rather than passing "first_block" and "last_block"
around, we now just pass a parent pointer.