Ensure that accelerator stream is passed to epub_init().
When epub_open_document() is passed a file stream, and an accelerator file
stream, it currently opens the file stream as an archive and passes it along
with the original file stream to epub_init(). The accelerator file stream is
ignored.

This is due to a typo in commit 2eb4c51a4c0ba5e57b4977890e02294d87e15685
which intented to pass the archive and the accelerator file stream to
epub_init().

This commit fixes the typo.