PDF text rendering; fix problems with T3 fonts.
The rules have changed slightly here for PDF 2.0. Acrobat
doesn't follow these rules. Broadly Chrome and Firefox do.

If a font is a Type 3, we have to modify the text rendering
mode if it's using clipping (i.e. if it's is larger than 3).

The current code was testing tos.text_mode, when this actually
lags behind the real value because of flushing. Accordingly,
test gstate->text.render instead.

Also, according to PDF 2.0, for modes 3 and 7, no type3 chars
should be drawn at all.