Bug 708895: Improve Type3 font bbox handling.
When we first run a type3 font we 'prepare' the glyphs;
i.e. we turn them all into display lists. At this point we
gather the rectangle used in the d1 operator (if any), and
use that for the char bbox.

For the supplied file, we have an incorrect font bbox [0 0 0 0],
but all the glyphs have valid d1 rectangles, despite being empty.

We have some code in here to detect completely empty glyphs, in
which case the glyph bbox is set to be "almost empty". This
effectively ignores the specified d1 rectangle, meaning that
selection rectangles are broken for this font.

This code was introduced back in October 2024 (commit
afaf3dc8801e863fdfafdda35192973fa635cb0e) to fix some timeouts.
The type3 mechanism has been extensively rewritten since then,
so this is no longer required. This if can be removed.

This leaves us with the simpler code starting with the 'else if'.
If the d1 rect is defined, we use it. If it's not, we calculate
it from the bbox of the glyph.

BUT... if we have the d1 rect, we also perform some additional
checks; if the font bbox is invalid (or the supplied d1 rectangle
does not fit in the font bbox), we distrust it, and recalculate it
from the actual glyph bbox.

So, for the given file we end up ignoring the d1 rectangle and
setting it to the empty bbox, which still breaks selection
rectangles.

The fix is to modify the code so that in the event of a completely
empty box, we'll take the d1 rect.

This opens us up to the possibility of a file having huge d1 rects
for empty chars. This would cause us to allocate lots of unnecessary
memory which would get cached.

Therefore we change the code so that, while we keep the 'correct'
d1 rect for each glyph, we drop the display list for empty ones.
Then, at render time, we can avoid allocating a large pixmap, and
instead just use a single pixel one.