Fix Type3 "cannot render glyph" warnings.
In the previous fix, I introduced an 'fz_t3_glyph_empty' function
to catch the empty-list case. Unfortunately, I then checked this
function just a few lines after the open-coded version of that
function was checked and we'd returned from the function.

Accordingly, I've now removed that function, and moved the
code up to the early exit position.

This correctly stops the warnings.