gl: Prevent OpenGL warnings about invalid values.
When the ui is initialized its cavity is the size of the window.
When ui element parcels are packed into the window the remaining
cavity shrinks.

A problem occurs when a parcel is larger than the window. At this
point the cavity coordinates can end up invalid, i.e. x0 >= x1
and y0 >= y1.

This eventually results in do_canvas() obtaining an invalid area
from ui_pack(), and uses it to set up a OpenGL scissor whose
dimensions end up negative, causing OpenGL to indicate an
GL_INVALID_VALUE error. This happens when opening a PDF with
small pages and opening the undo panel.

The fix is to clamp the modified ui.cavity coordinates in
ui_pack_layout() to the incoming ui.cavity coordinates, because
the cavity should never grow beyond its incoming dimensions.