Improve page segmentation.
Page segmentation works by making a list of rectangles where
content isn't on the page. It then looks for places where such
rectangles cut completely across the page (either from left to
right, or from top to bottom).

Previously, once we'd found the largest such subdivision, we'd
look at if it was a horizontal or vertical one. Based upon this
we'd slice the page repeatedly in that direction using all the
subdivisions we could find.

Unfortunately, this can produce bad results in several cases.

Imagine the following page layout:

  +-------------------+
  |                   |
  +-------------------+
  +-------------------+
  |                   |
  +-------------------+
                          <- A
  +--------+ +--------+
  |        | |        |
  +--------+ +--------+   /__ C
  +--------+ +--------+   \
  |        | |        |
  |        | |        |
  |        | +--------+
  +--------+ +--------+
  +--------+ |        |
  |        | |        |
  +--------+ +--------+

            ^
            |
            B

Here, the biggest gap is at A, the next biggest at B,
and there happens to be a small gap at C.

In the first pass, the existing code would fail to find
any vertical gaps at all, so would split at A and C.

What we really want to happen is for us to split at A
then B.

So this revised code now splits just once at the
largest gap before recursing to look for smaller ones.

Arguably this is still not quite ideal as in the case
where we have multiple equivalently sized gaps, we
might want to look at how 'sub splits' would work before
choosing which one to split with first. But that's
much more complex, and this is a definite step forward.

Also, update the debugging 'draw as postscript' code
to give more comprehensible results.