Table extraction: Introduce scoring for tables.
Calculate a 'score' for a table. This represents how
'awkward' a table looks.

Spanned cells are 'awkward'. Empty cells that don't
have borders are 'awkward'.

Too many such cells are an indication that maybe our
table is not right; perhaps a ghost table.

Throw away any tables that don't score below an
arbitrary threshold.

I expect this to be improved and refined in future,
but it's a reasonable first attempt.