Table extraction: Improve space handling within cells.
When marking the areas that are used within a table cell,
we were trying to mark single spaces as being used, and
to skip over multiple spaces.

This was being thwarted by mark_cells_for_content ignoring
any spaces at all, because the bbox for a space is empty.

Change the logic here to check for invalid rects rather
than empty ones, so the correct extent is still used.

This meant that synthetic spaces that cross table columns
are showing up unexpectedly. Make use of the new
FZ_STEXT_SYNTHETIC_LARGE character flag to mitigate this.