Remove soft-hyphens in the middle of lines when flattening stext.
Searching for "a" in these files

  * pymupdf/tests/resources/test_2861.pdf
  * pymupdf/tests/resources/test_3450.pdf
  * pymupdf/tests/resources/test_4702.pdf

results in an assertions triggered in debug builds:

  mutool: source/fitz/stext-search.c:1590: step_stext: Assertion `fz_is_unicode_whitespace(*pos)' failed.
  mutool: source/fitz/stext-search.c:1607: step_stext: Assertion `!fz_is_unicode_whitespace(c)' failed.

This happens because the spun text and structured text become
uncorrelated, which in turn happens because step_stext() and
do_flatten() do not agree about what happens to soft-hyphens in
the middle of a line of text: do_flatten() leaves them in the
text, and step_stext() assumes that they are removed.

This commit changes do_flatten() such that soft-hyphens are
removed just like step_stext() assumes.