Bug 708908: Respect AcroForm/NeedAppearances flag.
If this flag is set, we should resynth ALL widget appearances,
regardless of whether they already have an appearance.

Rejig needs_new_ap to indicate whether we need a new local (unsaved)
appearance in this case by setting it to a negative number.

Note: NeedAppearances is deprecated in PDF 2.0; but the 2.0 spec wording
claims this is because appearances are now required. That is slightly
at odds with the 1.7 spec that documents this flag as telling us to
create appearance streams for all widget annotations in the document
(not explicitly saying what to do if the widget already has an appearance).