Plug leak of softmask resources and colorspace.
When avoiding to apply softmasks to objects in the pattern by
dropping the softmask, it is also important to also drop the
references to the softmask resources and softmask colorspace.

The reason this is important is that in pdf_run_gs_SMask() the
softmask pointer is used to determine whether the resources and
colorspace should be dropped.

If the softmask has been dropped in pdf_show_pattern(), but the
references to the resources and colorspace remain then
pdf_run_gs_SMask() will not be able to drop them causing a leak.

This leak showed up in an ASAN build when rendering
tests_private/pdf/sumatra/x_-_renders_slowly.pdf.