Do not rely on a kept graphics state resource pointer being the same.
By changing pdf_keep_obj() so that it clones an object instead of
taking a new reference to one it is possible to more easily track
down memory leaks.

Doing this depends on the code not assuming that pdf_keep_obj()
returns the same pointer it received. So the calling code needs to
use the return value from pdf_keep_obj(). That's what changes here,
as well as setting graphics state pointers to NULL when dropping a
graphics state.

There is a chance that this approach falls down for some objects, but
if it helps pointing out leaks directly in some cases, it is is useful
to employ. Doing so requires applying the patches in this commit every
time, which is unnecessarily time-consuming, hence this commit.